# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GraduateWork",
        "version": "0.1.0"
    },
    "paths": {
        "/api/v1/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create",
                "operationId": "create_api_v1_user_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Remove",
                "operationId": "remove_api_v1_user_delete",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/api/v1/user/change_password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Change Password",
                "operationId": "change_password_api_v1_user_change_password_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordUserSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/api/v1/user/confirm_email": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Confirm Email",
                "operationId": "confirm_email_api_v1_user_confirm_email_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfirmEmailSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "login_api_v1_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenJWTSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "logout_api_v1_auth_logout_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh",
                "operationId": "refresh_api_v1_auth_refresh_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenJWTSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ChangePasswordUserSchema": {
                "properties": {
                    "current_password": {
                        "type": "string",
                        "minLength": 8,
                        "title": "Current Password",
                        "example": "securepassword123"
                    },
                    "new_password": {
                        "type": "string",
                        "minLength": 8,
                        "title": "New Password",
                        "example": "newsecurepassword123"
                    },
                    "confirm_new_password": {
                        "type": "string",
                        "minLength": 8,
                        "title": "Confirm New Password",
                        "example": "newsecurepassword123"
                    }
                },
                "type": "object",
                "required": [
                    "current_password",
                    "new_password",
                    "confirm_new_password"
                ],
                "title": "ChangePasswordUserSchema"
            },
            "ConfirmEmailSchema": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "otp_code": {
                        "type": "string",
                        "title": "Otp Code"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "otp_code"
                ],
                "title": "ConfirmEmailSchema"
            },
            "CreateUserSchema": {
                "properties": {
                    "username": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 3,
                        "title": "Username",
                        "example": "user123"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 8,
                        "title": "Password",
                        "example": "securepassword123"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email",
                        "example": "user@example.com"
                    },
                    "full_name": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 3,
                        "title": "Full Name",
                        "example": "Иванов Иван Иванович"
                    },
                    "phone_number": {
                        "type": "string",
                        "maxLength": 15,
                        "minLength": 10,
                        "title": "Phone Number",
                        "example": "+79990000000"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "password",
                    "email",
                    "full_name",
                    "phone_number"
                ],
                "title": "CreateUserSchema"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "LoginSchema": {
                "properties": {
                    "username": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 3,
                        "title": "Username",
                        "example": "user123"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 8,
                        "title": "Password",
                        "example": "securepassword123"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "password"
                ],
                "title": "LoginSchema"
            },
            "SuccessSchema": {
                "properties": {
                    "msg": {
                        "type": "string",
                        "title": "Msg",
                        "example": "Операция прошла успешно"
                    }
                },
                "type": "object",
                "required": [
                    "msg"
                ],
                "title": "SuccessSchema"
            },
            "TokenJWTSchema": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "title": "Access Token"
                    },
                    "refresh_token": {
                        "type": "string",
                        "title": "Refresh Token"
                    }
                },
                "type": "object",
                "required": [
                    "access_token",
                    "refresh_token"
                ],
                "title": "TokenJWTSchema"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "JWTBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}